Separate DataRequests namespace into Local and Deployed
**Why**: To help make it clearer which code runs locally vs on servers,
helps with separating autoloads

changelog: Internal, Source code, Reorganize data request classes to separate local vs deployed
