--- app/services/data_requests/lookup_user_by_uuid.rb
@@ -1,14 +1,0 @@
-module DataRequests
-  class LookupUserByUuid
-    attr_reader :uuid
-
-    def initialize(uuid)
-      @uuid = uuid
-    end
-
-    def call
-      User.find_by(uuid: uuid) ||
-        ServiceProviderIdentity.find_by(uuid: uuid)&.user
-    end
-  end
-end
