+++ lib/data_requests/deployed.rb
@@ -1,0 +1,6 @@
+require 'data_requests/deployed/create_email_addresses_report'
+require 'data_requests/deployed/create_mfa_configurations_report'
+require 'data_requests/deployed/create_user_events_report'
+require 'data_requests/deployed/create_user_report'
+require 'data_requests/deployed/lookup_shared_device_users'
+require 'data_requests/deployed/lookup_user_by_uuid'
