+++ lib/data_requests/local.rb
@@ -1,0 +1,4 @@
+require 'data_requests/local/fetch_cloudwatch_logs'
+require 'data_requests/local/write_cloudwatch_logs'
+require 'data_requests/local/write_user_events'
+require 'data_requests/local/write_user_info'
