+++ spec/lib/data_requests/deployed/create_email_addresses_report_spec.rb
@@ -1,6 +1,7 @@
 require 'rails_helper'
+require 'data_requests/deployed'
 
-describe DataRequests::CreateEmailAddressesReport do
+describe DataRequests::Deployed::CreateEmailAddressesReport do
   describe '#call' do
     it 'returns an array with hashes representing the users email addresses' do
       user = create(:user)
