+++ spec/lib/data_requests/deployed/create_mfa_configurations_report_spec.rb
@@ -1,6 +1,7 @@
 require 'rails_helper'
+require 'data_requests/deployed'
 
-describe DataRequests::CreateMfaConfigurationsReport do
+describe DataRequests::Deployed::CreateMfaConfigurationsReport do
   describe '#call' do
     it 'includes an array for phone numbers' do
       user = create(:user)
