+++ spec/lib/data_requests/deployed/create_user_events_report_spec.rb
@@ -1,6 +1,7 @@
 require 'rails_helper'
+require 'data_requests/deployed'
 
-describe DataRequests::CreateUserEventsReport do
+describe DataRequests::Deployed::CreateUserEventsReport do
   describe '#call' do
     it 'returns an array of hashes representing the users events' do
       user = create(:user)
