+++ spec/lib/data_requests/deployed/create_user_report_spec.rb
@@ -1,6 +1,7 @@
 require 'rails_helper'
+require 'data_requests/deployed'
 
-describe DataRequests::CreateUserReport do
+describe DataRequests::Deployed::CreateUserReport do
   it 'returns a report containing information about' do
     user = create(:user)
