+++ spec/lib/data_requests/deployed/lookup_shared_device_users_spec.rb
@@ -1,6 +1,7 @@
 require 'rails_helper'
+require 'data_requests/deployed'
 
-describe DataRequests::LookupSharedDeviceUsers do
+describe DataRequests::Deployed::LookupSharedDeviceUsers do
   describe '#call' do
     it 'recursively looks up users sharing devices' do
       user1 = create(:user)
