+++ spec/lib/data_requests/deployed/lookup_user_by_uuid_spec.rb
@@ -1,6 +1,7 @@
 require 'rails_helper'
+require 'data_requests/deployed'
 
-describe DataRequests::LookupUserByUuid do
+describe DataRequests::Deployed::LookupUserByUuid do
   describe '#call' do
     context 'when a user exists with the UUID' do
       it 'returns the user' do
