+++ spec/lib/data_requests/local/write_user_info_spec.rb
@@ -1,6 +1,7 @@
 require 'rails_helper'
+require 'data_requests/local'
 
-describe DataRequests::WriteUserInfo do
+describe DataRequests::Local::WriteUserInfo do
   describe '#call' do
     it 'writes a file with user information' do
       user_report = JSON.parse(
